+++ src/Symfony/Component/Serializer/CHANGELOG.md
@@ -6,6 +6,8 @@ CHANGELOG
 
  * deprecated the `XmlEncoder::TYPE_CASE_ATTRIBUTES` constant, use `XmlEncoder::TYPE_CAST_ATTRIBUTES` instead
  * added option to output a UTF-8 BOM in CSV encoder via `CsvEncoder::OUTPUT_UTF8_BOM_KEY` context option
+ * added the ability when denormalizing to cast values to the primitive type of their corresponding 
+   property via the `AbstractObjectNormalizer::CAST_PRIMITIVE_TYPES` context option
 
 4.3.0
 -----